// src/types/site-management.ts
export interface SiteConfig {
  siteName: string;
  siteDescription: string;
  siteUrl: string;
  logo: string;
  favicon: string;
  maintenanceMode: boolean;
  registrationEnabled: boolean;
  defaultLanguage: string;
  timezone: string;
  metaTitle: string;
  metaDescription: string;
  keywords: string;
  robotsTxt: string;
  facebookUrl: string;
  twitterUrl: string;
  whatsappUrl: string;
  telegramUrl: string;
  instagramUrl: string;
  cookieEnabled: boolean;
  cookieMessage: string;
  cookiePolicyUrl: string;
  headerScripts: string;
  footerScripts: string;
  customCss: string;
}